commits a bunch of things that I did tuesday like user auth and getting the map to load on the front page
