+++ client/public/css/styles.css
@@ -67,7 +67,7 @@ body {
   height: 400%;
   display: block;
   text-align: center;
-  z-index: 1
+  z-index: 1;
   position: relative;
 }
 
@@ -116,8 +116,8 @@ h1 {
 }*/
 
 #map {
-  width: 100%;
-  height: 60%;
+  width: 60vw;
+  height: 70vh;
   background-color: snow;
   /*border: 1px solid black;*/
 }
