+++ client/public/js/maps.js
@@ -1,0 +1,3 @@
+$(document).ready(function(){
+  console.log('maps');
+})
