+++ client/public/js/scripts.js
@@ -54,6 +54,30 @@
 // FORM VALIDATION NOT WORKING
 // $('.ui.form').form(validationRules, { onSuccess: submitForm });
 // END FORM VALIDATION NOT WORKING
+var map;
+function initMap() {
+  var myLatLng = {lat: parseFloat(williamsburg.latitude), lng: parseFloat(williamsburg.longitude)};
+  // Create a map object and specify the DOM element for display.
+  map = new google.maps.Map(document.getElementById('map'), {
+    center: myLatLng,
+    scrollwheel: false,
+    draggable: false,
+    zoom: 15
+  });
+}
+
+function createMarkers(){
+  for(var i=0; i< williamsburg.venues.length; i++){
+    if(williamsburg.venues[i].category == 'Bar'){
+      venueLoc = {lat: williamsburg.venues[i].lat, lng: williamsburg.venues[i].long }
+      var marker = new google.maps.Marker({
+        map: map,
+        position: venueLoc,
+        title: williamsburg.venues[i].name
+      });
+    }
+  }
+}
 
 // BEGINNING OF LATEST VERSION FORM VALIDATION
 $(document).ready(function() {
