+++ client/public/views/ajax.html
@@ -18,9 +18,6 @@
 
     <div id='map'></div>
 
-
-    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyAc9EXE72TMHf3k91hH-uRfKyaj8bftpko&callback=initMap"
-    async defer></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/2.2.1/jquery.min.js" charset="utf-8"></script>
     <script src="../js/ajax.js" charset="utf-8"></script>
   </body>
