+++ server/routes/index.js
@@ -1,5 +1,7 @@
 var express = require('express');
 var router = express.Router();
+
+var User = require('../models/user.js');
 /*
   'path' is needed because relative paths ../../ are considered malicious
   when importing modules in node. Example: importing routes in index.js
@@ -13,13 +15,16 @@ router.get('/', function(req, res, next) {
 
 
 router.get('/ajax', function(req, res, next){
+  var currentUser = JSON.parse(req.cookies.current_user);
+
+
   res.sendFile( path.resolve('client/public/views/ajax.html'));
+  // res.sendFile('public/index1.html' , { root : __dirname})
 })
 
 router.get('/login', function(req, res, next){
   res.sendFile( path.resolve('client/public/views/login_modal.html'));
 })
-
 
 router.get('/api/foursquare', function(req, res){
   res.json(data);
